Add files via upload
This is a action game which has a similar idea with a classic 2D scroll Mario games. I have not yet found a method to allow access from the C# to UNITY, so the current option is to 1: download the t18818my.unitypackage file or 2: download the t18818my.apk to a android device to play the game. The game itself is intended to be played with a device that has a touch screen, so playing on a computer with no touch screen is not recommended.